# inventory-be/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Automation and Tooling API",
    "description": "Comprehensive API documentation for the structural inventory platform. Includes Auth, RBAC, Users, Locations, Stocks, Requests, Audit, Notification, and Dashboard modules.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Resource not found."
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "example": 100 },
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 10 },
          "totalPages": { "type": "integer", "example": 10 }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "Jane Doe" },
          "username": { "type": "string", "example": "jane" },
          "role": { "type": "string", "example": "staff" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "parameters": {
      "PageParam": {
        "name": "page", "in": "query",
        "schema": { "type": "integer", "default": 1 }
      },
      "LimitParam": {
        "name": "limit", "in": "query",
        "schema": { "type": "integer", "default": 10 }
      },
      "SearchParam": {
        "name": "search", "in": "query",
        "schema": { "type": "string" }
      },
      "IdPathParam": {
        "name": "id", "in": "path", "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/users/profile": {
      "get": {
        "summary": "Get Current User Profile",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "Authenticated user's profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "data": { "$ref": "#/components/schemas/User" } }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized or user no longer exists",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        },
        "operationId": "getUsersProfile"
      }
    },
    "/users/profile/change-password": {
      "post": {
        "summary": "Self-service password change (auth only)",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password"],
                "properties": {
                  "password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "data": { "type": "object", "properties": { "message": { "type": "string" } } } }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        },
        "operationId": "postUsersProfileChangePassword"
      }
    },
    "/users": {
      "get": {
        "summary": "Get all users — paginated (Requires users:read)",
        "tags": ["Users"],
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "$ref": "#/components/parameters/SearchParam" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "getUsers"
      },
      "post": {
        "summary": "Create User directly — admin bootstrapping (Requires users:create)",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "username", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/User" } } }
              }
            }
          },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "postUsers"
      }
    },
    "/users/change-password": {
      "post": {
        "summary": "Admin force-change any user's password by username (Requires users:update)",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string" },
                        "user": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "username": { "type": "string" } } }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "User not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "postUsersChangePassword"
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get User by ID (Requires users:read)",
        "tags": ["Users"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": {
            "description": "User data",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/User" } } } } }
          },
          "404": { "description": "User not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "getUsersId"
      },
      "delete": {
        "summary": "Delete User (Requires users:delete)",
        "tags": ["Users"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": {
            "description": "Deleted user returned",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/User" } } } } }
          },
          "404": { "description": "User not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "deleteUsersId"
      }
    },
...
  }
}